scripts/csu_system_scripts/pyproject.toml
[project]
name = "csu-session-demo"
version = "0.1.0"
description = "Unified session management demo for CSU"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "pycryptodome>=3.20.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
